sensor/temperature/requirements.txt
birdseye==0.8.3
RPi.GPIO==0.7.0
paho_mqtt==1.5.0